More Logic Work
Added Rachel's portrait to the main scene.

Did more extensive work on the Bullet Hell logic which is:

Adding a crosshair that follows the cursor.
Clicking the crosshair on spawned enemies destroys them.
The enemies now explode/destroy on impact when touch with the player.

The only thing left here now is the timer and adding sprites. Please pull before you work.